Rename rebac pkg to rebac_admin, move backend setup to that pkg
